sql-rag-system/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pyodbc>=4.0.39
pandas>=2.0.0
sqlparse==0.4.4
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.4.2